viewing list of pdf names, navigating to specific pdf and displaying it all working
