+++ app/documents.tsx
@@ -1,38 +1,24 @@
+import { FileObject } from '@supabase/storage-js';
+import { Link } from 'expo-router';
 import { useEffect, useState } from 'react';
 import { Text, View } from 'react-native';
 
 import { LogoutButton } from '~/components/LogoutButton';
-import { PdfViewer } from '~/components/PdfViewer';
 import { useAuth } from '~/contexts/AuthContext';
 import { supabase } from '~/utils/supabase';
 
 export default function Details() {
   const { session } = useAuth();
-  const [pdfBlobText, setPdfBlobText] = useState<string | null>(null);
-
+  const [pdfs, setPdfs] = useState<FileObject[]>([]);
   useEffect(() => {
     if (!session) return;
     loadPdfs();
-    loadPdf();
   }, [session]);
 
   const loadPdfs = async () => {
     const { data } = await supabase.storage.from('documents').list(`${session?.user.id}`);
     if (!data) return;
-    console.log(
-      'list',
-      data.map((d) => d?.name)
-    );
-  };
-
-  const loadPdf = async () => {
-    const { data } = await supabase.storage
-      .from(`documents/${session?.user.id}`)
-      .download('Probate.pdf');
-    // .list(session?.user.id)
-    if (!data) return;
-    const blobUrl = URL.createObjectURL(data);
-    setPdfBlobText(blobUrl || null);
+    setPdfs(data);
   };
 
   return (
@@ -43,7 +29,11 @@ export default function Details() {
         <Text className={styles.subtitle}>
           Showing details for user {session?.user.email?.split('@')[0]}.
         </Text>
-        {pdfBlobText && <PdfViewer url={pdfBlobText} />}
+        {pdfs.map((pdf) => (
+          <Link key={pdf.id} href={`/document/${pdf.name}`}>
+            {pdf.name}
+          </Link>
+        ))}
         <LogoutButton />
       </View>
     </View>
